Custom ini support!
You can now have an ini that is separate from your Roms.ini! This is
good for working with multiple roms since you wont have to edit the
Roms.ini every time. All you need to do is have an ini with the same
name as the ROM in the same folder.
Example:
FireRed.gba
FireRed.ini