Pokemon Sprite Efficient Import
If the Efficient Import box is checked in pokemonedit, sprites will now be saved in Ram when they are imported. And if another sprite with the same imgstring or palstring is imported, it will not be inserted, instead it will just be repointed to the old image. Only stores the comparison info for as long as pokemonedit is open. So it will only work for pictures imported ina single session.